backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# File handling
python-multipart==0.0.6
aiofiles==23.2.1

# Vector database and RAG
chromadb==0.4.18
sentence-transformers==2.2.2

# AI/ML
google-generativeai==0.3.2
transformers==4.36.0
torch==2.1.1

# OCR and image processing
pytesseract==0.3.10
pillow==10.1.0
pdf2image==1.17.0
poppler-utils==0.1.0

# Excel processing
openpyxl==3.1.2
pandas==2.1.4

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment and configuration
python-dotenv==1.0.0

# HTTP client
httpx==0.25.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Additional utilities
python-slugify==8.0.1
uuid==1.30